main file renamed index.js => main.js
